requirements.txt
configparser==4.0.2
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.4
hexdump==3.3
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.6.1
more-itertools==5.0.0
numpy==1.23.4
packaging==21.3
Pillow==9.2.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-evtx==0.7.4
six==1.16.0
zipp==1.0.0
